v5.0.0-beta1 - SourceCode only

+++ ZimbraServer/src/java/com/zimbra/cs/db/DbOutOfOffice.java
@@ -37,6 +37,7 @@ import com.zimbra.common.util.Log;
 import com.zimbra.common.util.LogFactory;
 
 import com.zimbra.common.service.ServiceException;
+import com.zimbra.cs.account.AccountServiceException;
 import com.zimbra.cs.db.DbPool.Connection;
 import com.zimbra.cs.mailbox.Mailbox;
 
@@ -59,24 +60,17 @@ public class DbOutOfOffice {
      * @param cacheDurationMillis threshold for determining last sent time
      * @throws ServiceException if a database error occrurs
      */
-    public static boolean alreadySent(
-            Connection conn, Mailbox mbox, String sentTo, long cacheDurationMillis)
-    throws ServiceException
-    {
-        PreparedStatement stmt = null;
-        ResultSet rs = null;
+    public static boolean alreadySent(Connection conn, Mailbox mbox, String sentTo, long cacheDurationMillis)
+    throws ServiceException {
         sentTo = sentTo.toLowerCase();
         boolean result = false;
-        Timestamp cutoff = new Timestamp(
-                System.currentTimeMillis() - cacheDurationMillis);
-        
+        Timestamp cutoff = new Timestamp(System.currentTimeMillis() - cacheDurationMillis);
+
+        PreparedStatement stmt = null;
+        ResultSet rs = null;
         try {
-            stmt = conn.prepareStatement(
-                    "SELECT COUNT(*) " +
-                    "FROM " + TABLE_NAME + " " +
-                    "WHERE mailbox_id = ? " +
-                    "AND sent_to = ? " +
-                    "AND sent_on > ?");
+            stmt = conn.prepareStatement("SELECT COUNT(*) FROM " + TABLE_NAME +
+                    " WHERE mailbox_id = ? AND sent_to = ? AND sent_on > ?");
             stmt.setInt(1, mbox.getId());
             stmt.setString(2, sentTo);
             stmt.setTimestamp(3, cutoff);
@@ -86,8 +80,7 @@ public class DbOutOfOffice {
             int count = rs.getInt(1);
             result = (count > 0);
         } catch (SQLException e) {
-            throw ServiceException.FAILURE(
-                    "DbOutOfOffice.getSentTime: sql exception (mailbox_id="
+            throw ServiceException.FAILURE("DbOutOfOffice.getSentTime: sql exception (mailbox_id="
                             + mbox.getId() + " sent_to=" + sentTo + ")", e);
         } finally {
             DbPool.closeResults(rs);
@@ -126,37 +119,48 @@ public class DbOutOfOffice {
      * @param sentOn the timestamp of the out-of-office message
      * @throws ServiceException if a database error occurred
      */
-    public static void setSentTime(Connection conn, Mailbox mbox,
-            String sentTo, long sentOn) throws ServiceException {
+    public static void setSentTime(Connection conn, Mailbox mbox, String sentTo, long sentOn) throws ServiceException {
+        Timestamp ts = new Timestamp(sentOn);
+
         PreparedStatement stmt = null;
         try {
-            stmt = conn.prepareStatement(
-                    "INSERT INTO " + TABLE_NAME + "(mailbox_id, sent_to, sent_on) VALUES (?, ?, ?) " +
-                    "ON DUPLICATE KEY UPDATE sent_on = ?");
+            stmt = conn.prepareStatement("INSERT INTO " + TABLE_NAME + "(mailbox_id, sent_to, sent_on) VALUES (?, ?, ?) " +
+                    (Db.supports(Db.Capability.ON_DUPLICATE_KEY) ? "ON DUPLICATE KEY UPDATE sent_on = ?" : ""));
             stmt.setInt(1, mbox.getId());
-            stmt.setString(2, sentTo);
-            Timestamp ts = new Timestamp(sentOn);
+            stmt.setString(2, sentTo.toLowerCase());
             stmt.setTimestamp(3, ts);
-            stmt.setTimestamp(4, ts);
+            if (Db.supports(Db.Capability.ON_DUPLICATE_KEY))
+                stmt.setTimestamp(4, ts);
             int num = stmt.executeUpdate();
             if (num > 0) {
                 if (mLog.isDebugEnabled()) {
-                    mLog.debug("DbOutOfOffice.setSentTime: ok (mailbox_id="
-                            + mbox.getId() + " sent_to=" + sentTo + " sent_on="
-                            + sentOn + " rows=" + num + ")");
+                    mLog.debug("DbOutOfOffice.setSentTime: ok (mailbox_id=" + mbox.getId() +
+                            " sent_to=" + sentTo + " sent_on=" + sentOn + " rows=" + num + ")");
                 }
             } else {
-                mLog.error("DbOutOfOffice.setSentTime: no rows updated (mailbox_id="
-                        + mbox.getId()
-                        + " sent_to="
-                        + sentTo
-                        + " sent_on=" + sentOn + " rows=" + num + ")");
+                mLog.error("DbOutOfOffice.setSentTime: no rows updated (mailbox_id=" + mbox.getId()
+                        + " sent_to=" + sentTo + " sent_on=" + sentOn + " rows=" + num + ")");
             }
         } catch (SQLException e) {
-            throw ServiceException.FAILURE(
-                    "DbOutOfOffice.setSentTime: sql exception (mailbox_id="
-                    + mbox.getId() + " sent_to" + sentTo + " sent_on="
-                    + sentOn + ")", e);
+            if (!Db.supports(Db.Capability.ON_DUPLICATE_KEY) && Db.errorMatches(e, Db.Error.DUPLICATE_ROW)) {
+                try {
+                    stmt.close();
+
+                    stmt = conn.prepareStatement("UPDATE " + TABLE_NAME +
+                            " SET sent_on = ?" +
+                            " WHERE mailbox_id = ? AND sent_to = ?");
+                    stmt.setTimestamp(1, ts);
+                    stmt.setInt(2, mbox.getId());
+                    stmt.setString(3, sentTo.toLowerCase());
+                    stmt.executeUpdate();
+                } catch (SQLException nested) {
+                    throw ServiceException.FAILURE("DbOutOfOffice.setSentTime: sql exception " +
+                            "(mailbox_id=" + mbox.getId() + " sent_to" + sentTo + " sent_on=" + sentOn + ")", nested);
+                }
+            } else {
+                throw ServiceException.FAILURE("DbOutOfOffice.setSentTime: sql exception " +
+                        "(mailbox_id=" + mbox.getId() + " sent_to" + sentTo + " sent_on=" + sentOn + ")", e);
+            }
         } finally {
             DbPool.closeStatement(stmt);
         }
@@ -170,32 +174,37 @@ public class DbOutOfOffice {
      * @param mbox mailbox
      * @throws ServiceException if a database error occurred
      */
-    public static void clear(Connection conn, Mailbox mbox)
-            throws ServiceException {
+    public static void clear(Connection conn, String accountId) throws ServiceException {
         PreparedStatement stmt = null;
+        ResultSet rs = null;
         try {
+            stmt = conn.prepareStatement("SELECT id FROM zimbra.mailbox WHERE account_id = ?");
+            stmt.setString(1, accountId);
+            rs = stmt.executeQuery();
+            if (!rs.next())
+                throw AccountServiceException.NO_SUCH_ACCOUNT(accountId);
+            int mailboxId = rs.getInt(1);
+            rs.close();
+            stmt.close();
+
             stmt = conn.prepareStatement("DELETE FROM " + TABLE_NAME + " WHERE mailbox_id = ?");
-            stmt.setInt(1, mbox.getId());
+            stmt.setInt(1, mailboxId);
             int num = stmt.executeUpdate();
-            mLog.debug("DbOutOfOffice.clear() mbox=" + mbox + " rows=" + num);
+            mLog.debug("DbOutOfOffice.clear() mbox=" + mailboxId + " rows=" + num);
         } catch (SQLException e) {
-            throw ServiceException.FAILURE("DbOutOfOffice.clear mbox="
-                    + mbox.getId(), e);
+            throw ServiceException.FAILURE("DbOutOfOffice.clear acctId=" + accountId, e);
         } finally {
+            DbPool.closeResults(rs);
             DbPool.closeStatement(stmt);
         }
     }
     
     public static void prune(Connection conn, long cacheDurationMillis)
-    throws ServiceException
-    {
+    throws ServiceException {
         PreparedStatement stmt = null;
         try {
-            Timestamp cutoff = new Timestamp(
-                    System.currentTimeMillis() - cacheDurationMillis);
-            stmt = conn.prepareStatement(
-                    "DELETE FROM " + TABLE_NAME + " " +
-                    "WHERE sent_on <= ?");
+            Timestamp cutoff = new Timestamp(System.currentTimeMillis() - cacheDurationMillis);
+            stmt = conn.prepareStatement("DELETE FROM " + TABLE_NAME + " WHERE sent_on <= ?");
             stmt.setTimestamp(1, cutoff);
             int num = stmt.executeUpdate();
             if (mLog.isDebugEnabled()) {
